+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 8.2.0-dev
 - allowed skipping test when env variable is not set (to a value)
 - allowed not loading and running test suite from certain subfolders
 - added option to retry a test on failure
+- allowed specifying operator for attribute RequiresPhpVersion
 
 Version 8.1.1
 - fixed error in mytester.php when package konecnyjakub/phpt-runner is not installed